refactor: further cleanup
