--- src/customIcons.js
@@ -1,25 +1,0 @@
-/*global UIkit*/
-UIkit.icon.add(
-  'question-plain',
-  `<svg width="20" height="20" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><circle cx="10.44" cy="14.42" r="1.05"/><path fill="none" stroke="#000" stroke-width="1.2" d="M8.17,7.79 C8.17,4.75 12.72,4.73 12.72,7.72 C12.72,8.67 11.81,9.15 11.23,9.75 C10.75,10.24 10.51,10.73 10.45,11.4 C10.44,11.53 10.43,11.64 10.43,11.75"/></svg>`,
-);
-UIkit.icon.add('blank', `<svg width="20" height="20" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"></svg>`);
-UIkit.icon.add(
-  'chart-line',
-  `<svg width="20" height="20" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512"><path d="M496 384H64V80c0-8.84-7.16-16-16-16H16C7.16 64 0 71.16 0 80v336c0 17.67 14.33 32 32 32h464c8.84 0 16-7.16 16-16v-32c0-8.84-7.16-16-16-16zM464 96H345.94c-21.38 0-32.09 25.85-16.97 40.97l32.4 32.4L288 242.75l-73.37-73.37c-12.5-12.5-32.76-12.5-45.25 0l-68.69 68.69c-6.25 6.25-6.25 16.38 0 22.63l22.62 22.62c6.25 6.25 16.38 6.25 22.63 0L192 237.25l73.37 73.37c12.5 12.5 32.76 12.5 45.25 0l96-96 32.4 32.4c15.12 15.12 40.97 4.41 40.97-16.97V112c.01-8.84-7.15-16-15.99-16z"></path></svg>`,
-);
-UIkit.icon.add(
-  'chart-line-ban',
-  `<svg width="20" height="20" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20">
-    <path d="M 16.340042,13.621398 H 4.927966 V 5.5906772 c 0,-0.2335245 -0.1891446,-0.4226692 -0.4226695,-0.4226692 h -0.845339 c -0.2335249,0 -0.4226695,0.1891447 -0.4226695,0.4226692 v 8.8760598 c 0,0.466786 0.3785534,0.845339 0.845339,0.845339 h 12.257415 c 0.233525,0 0.42267,-0.189145 0.42267,-0.42267 v -0.845339 c 0,-0.233524 -0.189145,-0.422669 -0.42267,-0.422669 z M 15.494703,6.013347 h -3.118772 c -0.564792,0 -0.847717,0.6828754 -0.448294,1.0822979 L 12.783542,7.95155 10.845339,9.890019 8.907135,7.951815 c -0.33021,-0.3302105 -0.865416,-0.3302105 -1.1953621,0 L 5.8972,9.766388 c -0.1651053,0.165105 -0.1651053,0.432708 0,0.597813 l 0.5975489,0.597549 c 0.1651054,0.165106 0.4327079,0.165106 0.5978131,0 L 8.3093219,9.744726 10.247525,11.68293 c 0.330211,0.330211 0.865416,0.330211 1.195363,0 l 2.536017,-2.536017 0.855905,0.855905 c 0.399423,0.399423 1.082298,0.116499 1.082298,-0.448293 V 6.4360162 C 15.917373,6.2024917 15.728228,6.013347 15.494703,6.013347 Z" />
-    <circle fill="none" stroke="#000" stroke-width="1.1" cx="10" cy="10" r="9"/><line fill="none" stroke="#000" stroke-width="1.1" x1="4" y1="3.5" x2="16" y2="16.5"/>
-  </svg>`,
-);
-UIkit.icon.add(
-  'pin',
-  `<svg width="20" height="20" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512" class="svelte-c8tyih"><path d="M256 48c-79.5 0-144 59.9-144 133.7 0 104 144 282.3 144 282.3s144-178.3 144-282.3C400 107.9 335.5 48 256 48zm0 190.9c-25.9 0-46.9-21-46.9-46.9s21-46.9 46.9-46.9 46.9 21 46.9 46.9-21 46.9-46.9 46.9z"></path></svg>`,
-);
-UIkit.icon.add(
-  'chevron-right-end',
-  `<svg width="20" height="20" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><path fill="none" stroke="#000" stroke-width="1.03" d="M6 4 L12 10 L6 16 M14 4 L14 16"></path></svg>`,
-);
