+++ src/styles/uikit.css
@@ -1,41 +1,3 @@
-/*
- * 1. Using `flex` instead of `inline-block` to prevent whitespace between child elements
- * 2. Behave like button
- * 3. Create position context
- */
-.uk-vertical-button-group {
-  /* 1 */
-  display: inline-flex;
-  flex-direction: column;
-  /* 2 */
-  vertical-align: middle;
-  /* 3 */
-  position: relative;
-}
-
-/* Group
-  ========================================================================== */
-/*
-  * Collapse border
-  */
-.uk-vertical-button-group > .uk-button:nth-child(n + 2),
-.uk-vertical-button-group > div:nth-child(n + 2) .uk-button {
-  margin-top: -1px;
-}
-
-.uk-vertical-button-group .uk-button:hover,
-.uk-vertical-button-group .uk-button:focus,
-.uk-vertical-button-group .uk-button:active,
-.uk-vertical-button-group .uk-button.uk-active {
-  position: relative;
-  z-index: 1;
-}
-
-.uk-icon-button-small {
-  width: 20px;
-  height: 20px;
-}
-
 .uk-button-delphi {
   background: #121232;
   color: white;
