+++ src/theme.js
@@ -2,13 +2,6 @@ export const MISSING_COLOR = '#eeeeee';
 export const ZERO_COLOR = 'rgb(242,242,242)';
 
 export const MAP_THEME = {
-  selectedRegionOutline: '#333333', //'#000000',
-  hoverRegionOutline: '#cccccc', //'#ffffff',
-  stateFill: 'rgba(0, 0, 0, 0)',
-  stateOutline: '#ffffff', //'#bcbcbc',
-  countyFill: '#eeeeee', //'#e4dac4', // missing value
-  countyOutline: '#ffffff', //'#e0e0e0',
-  countyOutlineWhenFilled: '#ffffff', //'#616161',
+  hoverRegionOutline: '#cccccc',
+  stateOutline: '#ffffff',
 };
-
-export const selectionColors = ['#0F0589', '#16A997', '#A71808', '#CD22B2', '#F58B29'];
